Improve performance of create_faces()
By not using a dictionary, this function becomes faster by about a
factor of 10. First of all, we do not need to go through all cells, but
only the inner cells. There are not faces between outer cells, so this
is fine. Then, we can just skip every connection, where the left index
is larger than the right index. This works because the cell connection
graph is symmetric so we are guaranteed to come by every connection
eventually. After gather all the face information in an array, we just
need to unpack it into the mesh structure.