Improve performance of compute_send_graph()
We still use a kdtree because we need to be able to compare with a
tolerance, but we only insert absolutely necessary centers. If the mesh
was created (mesh_create()) then we need to match on the center of the
outer cell (for periodic and neighbor cells), but insert the index of
the connected cell (there is only ever one). If the mesh was read
(mesh_read()) then for neighbor cells, we need to match on the center of
the inner cell and still insert the index of the connected cell. Because
at this point (mesh_commit()) we do not know if the mesh was created or
read, we just insert both, but crucially, we do not insert every inner
cell. This reduces the size of the resulting kdtree drastically and
improves the performance of this function by about a factor 10.