Remove strbuf, add Name type for names, fix HDF5 IO for strings
The strbuf type was starting to annoy me, so I removed it. The Name type
does what I want and is less cumbersome to use. I add a way to pass the
size of the strings to during HDF5 IO so that the API is more flexible.
For the most part, everything stays the same, but it's slightly nicer
now.