Add equations module
An equation system can track scalars, vectors, or matrices. Types are
defined at creation. Equation system properties are scalar. For the most
part, the equation system keeps track of function pointers, which
compute things like timestep, boundary conditions, or fluxes. Gradient
limiters are part of the equations module and can be selected at
creation. A user may chose to add number of user variables to the
equation system, these are only computed during write and do not
directly effect the equation system. On chose of user variables are, for
example, an exact solution. A user should use the create and set
functions to modify the equation system, as these contain plausibility
checks and perform the setup work, if necessary. For the most part,
things are selected via strings, so be careful with spelling. If you are
looking up something (maybe a boundary condition or an entity) and it
cannot be found, this will throw an error, you you will surely notice.