Bugfix: do not rotate conserved variables
For first order in space, you can just rotate the momentum vector. For
second order in space, it is wrong to rotate the momentum vector,
because we are getting reconstructed values which do not necessarily
produce correct conserved variables. I think this is due to the fact the
momentum is the product of two variables and computing the gradient of
it should involve a chain rule. Anyways, for the global_to_local()
functions, we now only rotate the primitive variables and recompute the
conserved variables.

I should also be faster to compute the conserved variables in that way,
because it involves less operations. I did not benchmark this properly.

For the global_to_local() in boundary.c we can get away with an even
simpler form, since we never pass reconstructed values to the function.