Friendship ended with number
Using ptrdiff_t through number was a good idea. However, I just don't
like it anymore. MPI_INT is about 2 billion large. An array of 2 billion
doubles would be around 16GB and a full mesh with 2 billion cells would
probably be around over 100GB in total. If I ever get to do simulations
with such large meshes (X for doubt...) I will rethink this decision.
For now, I will switch to int and use this everywhere as the default
integer. The only exception is the arena. If capacity is an int, then
the largest arena you could create would be 2GB, which is too small.
Therefore, I use ptrdiff_t directly for sizes in of arena functions.