Replace type array with size array (names subject to change)
Using an enum for the sizes of the variables is a bit cumbersome and has
some annoying side effects. I switched it to an integer array of sizes
which represent the sizes of the variables. I am not a big fan of the
name 'size' yet, so this might change in the future.