Fix output format
In order for ParaView to successfully show partitioned output file, the
ghost/periodic/neighbor cells need to be included in the output data.
They also need to be marked with the correct vtkGhostType value:

- 0: inner cell
- EXTERIORCELL | HIDDENCELL: ghost/periodic cell
- DUPLICATECELL: neighbor cell

If we do this, the output is correct, and filter such as iso-surface of
gradient work as expected.