Fix integration of fluxes
Viscous fluxes should use unlimited gradients. I changed the O2 function
order as follows:

1. Compute gradients
2. Compute viscous fluxes
3. Limit gradients
4. Compute convective fluxes

That way the viscous fluxes use unlimited gradients. There was also a
small bug in how the average gradient for ghost cells was computed. That
is fixed now.

The unfortunate effect of this change is that we have to add one more
round of gradient synchronization. At first though it might seem
unnecessary the synchronize the gradients for the convective fluxes
again, since they have already been synchronized for the viscous fluxes.
However, if the limiter is applied, the inner gradients might change, so
we need to re-synchronize. It is only unnecessary to synchronize, if we
running viscous and convective fluxes, without limiting. However, I
would say that this an unlikely simulation setup, so I will not optimize
for it.