Add mesh write variant for single partition
At this moment, I don't know how to properly write partitioned dataset
using VTKHDF, but I am working on finding a [solution][1]. This commit
adds the previous writer back to the code as mesh_write1(). This variant
creates a mesh file which looks like it has only a single partition. In
that way, filters such as Gradient of Contour work as expected. No gaps.

However, this mesh format is not infinitely scalable and ParaView cannot
open the output files if they are very large. I don't know the exact
cutoff size. Also, I suspect that if I write output like this, ParaView
will not be able to parallelize the visualization. If you are working
with large datasets, you can switch to the partitioned mesh writer
mesh_write2() in src/mesh.h by modifying the #define. The output that
this produces can be processed in parallel by ParaView, but there will
be gaps in this like iso-surfaces.

In the future I might switch to [CGNS](https://cgns.org/index.html) for
storing outputs, as I would expect this format to handle partitioned
datasets better than VTKHDF. But let's see what the discussion on the
ParaView forum produces...

[1]: https://discourse.paraview.org/t/vtkhdf-partitioned-dataset-visualization-issues/17353/8