Write results into file as one partition
This will work for the moment. ParaView can visualize it without any
issues. With the entity array we can separate out surfaces, even though
the boundary cells are actually not the surface cells. However, since
the cells are flat, this still works.

If the mesh gets large (I don't know exactly how large) we will not be
able to open it anymore. At this point, we will need a better solution
for the output. At moment I think I will write a script/program that
takes the single-partition mesh file, and partitions it for
visualization.